+++ octilab/actuators/__init__.py
@@ -22,7 +22,7 @@ which defines the common interface for all actuator models. The actuator models 
 and called by the :class:`omni.isaac.lab.assets.Articulation` class.
 """
 
-from .actuator_cfg import HebiMotorCfg
-from .actuator_pd import HebiMotor
+from .actuator_cfg import HebiStrategy3ActuatorCfg, EffortMotorCfg, HebiStrategy4ActuatorCfg, HebiDCMotorCfg
+from .actuator_pd import HebiStrategy3Actuator, EffortMotor, HebiStrategy4Actuator, HebiDCMotor
 
-__all__ = [HebiMotorCfg, HebiMotor]
+__all__ = [HebiStrategy3Actuator, HebiStrategy3Actuator, HebiStrategy4Actuator]