+++ octilab/actuators/actuator_pd.py
@@ -14,67 +14,142 @@ from omni.isaac.core.utils.types import ArticulationActions
 from omni.isaac.lab.actuators.actuator_base import ActuatorBase
 
 if TYPE_CHECKING:
-    from .actuator_cfg import HebiMotorCfg
-
-from octilab.controllers.TychoController import TychoController
+    from .actuator_cfg import HebiStrategy3ActuatorCfg, EffortMotorCfg, HebiStrategy4ActuatorCfg, HebiDCMotorCfg
 
 
-class HebiMotor(ActuatorBase):
-    r"""
-    Direct control (DC) motor actuator model with velocity-based saturation model.
+class EffortMotor(ActuatorBase):
+    cfg: EffortMotorCfg
 
-    It uses the same model as the :class:`IdealActuator` for computing the torques from input commands.
-    However, it implements a saturation model defined by DC motor characteristics.
+    def __init__(self, cfg: EffortMotorCfg, *args, **kwargs):
+        super().__init__(cfg, *args, **kwargs)
 
-    A DC motor is a type of electric motor that is powered by direct current electricity. In most cases,
-    the motor is connected to a constant source of voltage supply, and the current is controlled by a rheostat.
-    Depending on various design factors such as windings and materials, the motor can draw a limited maximum power
-    from the electronic source, which limits the produced motor torque and speed.
+        self.actuation_limit = torch.tensor(cfg.actuation_limit, device=self._device)
 
-    A DC motor characteristics are defined by the following parameters:
+    def compute(self, control_action: ArticulationActions, joint_pos: torch.Tensor, joint_vel: torch.Tensor):
+        self.computed_effort = control_action.joint_efforts * self.actuation_limit
+        self.applied_effort = self.computed_effort
+        control_action.joint_efforts = self.applied_effort
+        control_action.joint_positions = None
+        control_action.joint_velocities = None
+        return control_action
 
-    * Continuous-rated speed (:math:`\dot{q}_{motor, max}`) : The maximum-rated speed of the motor.
-    * Continuous-stall torque (:math:`\tau_{motor, max}`): The maximum-rated torque produced at 0 speed.
-    * Saturation torque (:math:`\tau_{motor, sat}`): The maximum torque that can be outputted for a short period.
+    def reset(self, env_ids: Sequence[int]):
+        pass
 
-    Based on these parameters, the instantaneous minimum and maximum torques are defined as follows:
 
-    .. math::
+class HebiStrategy3Actuator(ActuatorBase):
 
-        \tau_{j, max}(\dot{q}) & = clip \left (\tau_{j, sat} \times \left(1 -
-            \frac{\dot{q}}{\dot{q}_{j, max}}\right), 0.0, \tau_{j, max} \right) \\
-        \tau_{j, min}(\dot{q}) & = clip \left (\tau_{j, sat} \times \left( -1 -
-            \frac{\dot{q}}{\dot{q}_{j, max}}\right), - \tau_{j, max}, 0.0 \right)
-
-    where :math:`\gamma` is the gear ratio of the gear box connecting the motor and the actuated joint ends,
-    :math:`\dot{q}_{j, max} = \gamma^{-1} \times  \dot{q}_{motor, max}`, :math:`\tau_{j, max} =
-    \gamma \times \tau_{motor, max}` and :math:`\tau_{j, peak} = \gamma \times \tau_{motor, peak}`
-    are the maximum joint velocity, maximum joint torque and peak torque, respectively. These parameters
-    are read from the configuration instance passed to the class.
-
-    Using these values, the computed torques are clipped to the minimum and maximum values based on the
-    instantaneous joint velocity:
-
-    .. math::
-
-        \tau_{j, applied} = clip(\tau_{computed}, \tau_{j, min}(\dot{q}), \tau_{j, max}(\dot{q}))
-
-    """
-
-    cfg: HebiMotorCfg
+    cfg: HebiStrategy3ActuatorCfg
     """The configuration for the actuator model."""
 
-    def __init__(self, cfg: HebiMotorCfg, *args, **kwargs):
+    def __init__(self, cfg: HebiStrategy3ActuatorCfg, *args, **kwargs):
         super().__init__(cfg, *args, **kwargs)
-        # parse configuration
-        if self.cfg.gain_xml_path is None:
-            raise ValueError("To use HebiMotor you must specify the gain file in HebiMotorCfg")
+        self.num_envs = kwargs['num_envs']
+        self.device = kwargs['device']
+        self.kp = torch.tensor(cfg.kp, device=self.device).view(-1)
+        self.ki = torch.tensor(cfg.ki, device=self.device).view(-1)
+        self.kd = torch.tensor(cfg.kd, device=self.device).view(-1)
+        self.i_clamp = torch.tensor(cfg.i_clamp, device=self.device).view(-1)
+        self.min_target = torch.tensor(cfg.min_target, device=self.device).view(-1)
+        self.max_target = torch.tensor(cfg.max_target, device=self.device).view(-1)
+        self.target_lowpass = torch.tensor(cfg.target_lowpass, device=self.device).view(-1)
+        self.min_output = torch.tensor(cfg.min_output, device=self.device).view(-1)
+        self.max_output = torch.tensor(cfg.max_output, device=self.device).view(-1)
+        self.output_lowpass = torch.tensor(cfg.output_lowpass, device=self.device).view(-1)
+        self.d_on_error = torch.tensor(cfg.d_on_error, device=self.device).view(-1)
+        self.maxtorque = torch.tensor(cfg.maxtorque, device=self.device).view(-1)
+        self.speed_24v = torch.tensor(cfg.speed_24v, device=self.device).view(-1)
 
-        if self.cfg.actuator_biasprm is not None:
-            self.actuator_biasprm = torch.tensor(self.cfg.actuator_biasprm, device=self._device)
-        else:
-            raise ValueError("To use HebiMotor you must specify the actuator_biasprm in HebiMotorCfg")
-        self.ctrl = TychoController(self.cfg.gain_xml_path, self.cfg.only_position_control)
+        self.errSum = torch.zeros((self.num_envs, 21), device=self.device)
+        self.lastDOn = torch.zeros((self.num_envs, 21), device=self.device)
+        self.last_target = torch.zeros((self.num_envs, 21), device=self.device)
+        self.last_output = torch.zeros((self.num_envs, 21), device=self.device)
+
+    """
+    Operations.
+    """
+    def compute(
+        self,
+        control_action: ArticulationActions,
+        joint_pos: torch.Tensor,
+        joint_vel: torch.Tensor,
+        joint_effort: torch.Tensor
+    ) -> ArticulationActions:
+        joint_des = torch.cat([control_action.joint_positions,
+                               control_action.joint_velocities,
+                               control_action.joint_efforts], dim=1)
+        joint = torch.cat([joint_pos, joint_vel, joint_effort], dim=1)
+        joint_des = torch.clip(joint_des, self.min_target, self.max_target)
+        # broadcasting rule is applied at | here
+        target = torch.where((self.last_target == 0) | (self.target_lowpass == 1),
+                             joint_des, self.last_target + self.target_lowpass * (joint_des - self.last_target))
+        self.last_target = target
+        error = target - joint
+        self.errSum += error
+
+        dErr = torch.where(self.lastDOn == 1, error - self.lastDOn, self.lastDOn - joint)
+        self.lastDOn = torch.where(self.lastDOn == 1, error, joint)
+        i_output = torch.clip(self.errSum * self.ki, -self.i_clamp, self.i_clamp)
+        output = self.kp * error + i_output + self.kd * dErr
+        output = torch.clip(output, self.min_output, self.max_output)
+        # broadcasting rule is applied at | here
+        output = torch.where((self.output_lowpass == 1),
+                             output, self.last_output + self.output_lowpass * (output - self.last_output))
+        self.last_output = output
+        pwm = torch.sum(output.view(self.num_envs, 3, -1), dim=1)
+        torque_des = self.pwm_to_torque_(pwm, joint_vel)
+
+        # setting below field passed to "data" for record purpose
+        self.computed_effort = torque_des
+        self.applied_effort = torque_des
+
+        # control_action.joint_efforts is the field will be passed to simulator for physic stepping
+        control_action.joint_efforts = self.applied_effort
+        control_action.joint_positions = None
+        control_action.joint_velocities = None
+        return control_action
+
+    def pwm_to_torque_(self, pwm, joint_vel):
+        ctrl = torch.multiply(pwm - torch.divide(torch.abs(joint_vel), self.speed_24v), self.maxtorque)
+        ctrl = torch.clip(ctrl, -self.maxtorque, self.maxtorque)
+        return ctrl
+
+    def reset(self, env_ids: Sequence[int]):
+        self.errSum[env_ids] = 0
+        self.lastDOn[env_ids] = 0
+        self.last_target[env_ids] = 0
+        self.last_output[env_ids] = 0
+
+
+class HebiStrategy4Actuator(ActuatorBase):
+
+    cfg: HebiStrategy4ActuatorCfg
+    """The configuration for the actuator model."""
+
+    def __init__(self, cfg: HebiStrategy4ActuatorCfg, *args, **kwargs):
+        super().__init__(cfg, *args, **kwargs)
+
+        self.num_envs = kwargs['num_envs']
+        self.device = kwargs['device']
+        self.p_p = torch.tensor(cfg.p_p, device=self.device).view(-1)
+        self.p_d = torch.tensor(cfg.p_d, device=self.device).view(-1)
+        self.e_p = torch.tensor(cfg.e_p, device=self.device).view(-1)
+        self.e_d = torch.tensor(cfg.e_d, device=self.device).view(-1)
+        self.i_clamp = torch.tensor(cfg.i_clamp, device=self.device).view(-1)
+        self.min_target = torch.tensor(cfg.min_target, device=self.device).view(-1)
+        self.max_target = torch.tensor(cfg.max_target, device=self.device).view(-1)
+        self.target_lowpass = torch.tensor(cfg.target_lowpass, device=self.device).view(-1)
+        self.min_output = torch.tensor(cfg.min_output, device=self.device).view(-1)
+        self.max_output = torch.tensor(cfg.max_output, device=self.device).view(-1)
+        self.output_lowpass = torch.tensor(cfg.output_lowpass, device=self.device).view(-1)
+        self.d_on_error = torch.tensor(cfg.d_on_error, device=self.device).view(-1)
+        self.maxtorque = torch.tensor(cfg.maxtorque, device=self.device).view(-1)
+        self.speed_24v = torch.tensor(cfg.speed_24v, device=self.device).view(-1)
+        self.errSum = torch.zeros((self.num_envs, 21), device=self.device)
+        self.lastDOn = torch.zeros((self.num_envs, 21), device=self.device)
+        self.last_target = torch.zeros((self.num_envs, 21), device=self.device)
+        self.last_output = torch.zeros((self.num_envs, 21), device=self.device)
+        self.last_effort = torch.zeros((self.num_envs, 7), device=self.device)
 
     """
     Operations.
@@ -89,19 +164,17 @@ class HebiMotor(ActuatorBase):
         if control_action.joint_positions is None:
             raise ValueError("Hebi pwm motor requires a desired(target) joint position to calculate the \
                             the joint force. You must use action that outputs joint position.")
-        elif len(control_action.joint_positions) > 1:
-            raise ValueError("Hebi pwm motor does not support vectorized environment processing, \
-                             please set num_env to 1 if you want to use HebiMotor")
+
         else:
-            pwm = self.ctrl.act(
-                control_action.joint_positions[0].tolist(),
-                None,
-                joint_pos[0].cpu().numpy(),
-                joint_vel[0].cpu().numpy(),
-                joint_effort[0].cpu().numpy(),
-                )
-            pwm = torch.tensor(pwm, device=self._device)
-            torque_des = self._pwm_to_torque(pwm, self.actuator_biasprm, joint_vel,)
+            error_pos = control_action.joint_positions - joint_pos
+            target_eff = self.p_p * error_pos + self.p_d * joint_vel + control_action.joint_efforts
+
+            error_eff = target_eff - joint_effort
+            impulse = target_eff - self.last_effort
+            pwm = self.e_p * error_eff + self.e_d * impulse
+            pwm = torch.clip(pwm, -1, 1)
+            torque_des = self.pwm_to_torque_(target_eff, joint_vel)
+            self.last_effort[:] = torque_des
 
             # setting below field passed to "data" for record purpose
             self.computed_effort = torque_des
@@ -114,18 +187,70 @@ class HebiMotor(ActuatorBase):
             return control_action
 
     def reset(self, env_ids: Sequence[int]):
-        pass
+        self.errSum[env_ids] = 0
+        self.lastDOn[env_ids] = 0
+        self.last_target[env_ids] = 0
+        self.last_output[env_ids] = 0
+
+    def pwm_to_torque_(self, pwm, joint_vel):
+        ctrl = torch.multiply(pwm - torch.divide(torch.abs(joint_vel), self.speed_24v), self.maxtorque)
+        ctrl = torch.clip(ctrl, -self.maxtorque, self.maxtorque)
+        return ctrl
+
+
+class HebiDCMotor(ActuatorBase):
+
+    cfg: HebiDCMotorCfg
+    """The configuration for the actuator model."""
+
+    def __init__(self, cfg: HebiDCMotorCfg, *args, **kwargs):
+        super().__init__(cfg, *args, **kwargs)
+
+        self.p_p = torch.tensor(cfg.p_p, device=self._device).view(-1)
+        self.p_d = torch.tensor(cfg.p_d, device=self._device).view(-1)
+        self.e_p = torch.tensor(cfg.e_p, device=self._device).view(-1)
+        self.e_d = torch.tensor(cfg.e_d, device=self._device).view(-1)
+        self.speed_24v = torch.tensor(cfg.speed_24v, device=self._device).view(-1)
+        self.saturation_effort = torch.tensor(cfg.saturation_effort, device=self._device).view(-1)
+        self.maxtorque = torch.tensor(cfg.maxtorque, device=self._device).view(-1)
+        # prepare joint vel buffer for max effort computation
+        self._joint_vel = torch.zeros_like(self.computed_effort)
+        # create buffer for zeros effort
+        self._zeros_effort = torch.zeros_like(self.computed_effort)
+
+    """
+    Operations.
+    """
+
+    def compute(
+        self, control_action: ArticulationActions, joint_pos: torch.Tensor, joint_vel: torch.Tensor
+    ) -> ArticulationActions:
+        error_pos = control_action.joint_positions - joint_pos
+        # calculate the desired joint torques
+        self.computed_effort = self.p_p * 5 * error_pos + self.p_d * joint_vel + control_action.joint_efforts
+        # clip the torques based on the motor limits
+        self.applied_effort = self._clip_effort(self.computed_effort)
+        # set the computed actions back into the control action
+        control_action.joint_efforts = self.applied_effort
+        control_action.joint_positions = None
+        control_action.joint_velocities = None
+        return control_action
 
     """
     Helper functions.
     """
-    def _pwm_to_torque(self, pwm: torch.Tensor, biasprm: torch.Tensor, joint_vel, gravity=None):
-        pwm = torch.clip(pwm, -1, 1)
-        maxtorque = biasprm[0][0:7]
-        speed_24v = biasprm[1][0:7]
-        qvel = joint_vel
-        ctrl = torch.multiply(pwm - torch.divide(torch.abs(qvel), speed_24v), maxtorque)
-        if gravity is not None:
-            ctrl += gravity
-        ctrl = torch.clip(ctrl, -maxtorque, maxtorque)
-        return ctrl
+
+    def _clip_effort(self, effort: torch.Tensor) -> torch.Tensor:
+        # compute torque limits
+        # -- max limit
+        max_effort = self.saturation_effort * (1.0 - self._joint_vel / self.speed_24v)
+        max_effort = torch.clip(max_effort, min=self._zeros_effort, max=self.maxtorque)
+        # -- min limit
+        min_effort = self.saturation_effort * (-1.0 - self._joint_vel / self.speed_24v)
+        min_effort = torch.clip(min_effort, min=-self.maxtorque, max=self._zeros_effort)
+
+        # clip the torques based on the motor limits
+        return torch.clip(effort, min=min_effort, max=max_effort)
+
+    def reset(self, env_ids: Sequence[int]):
+        pass