+++ octilab/envs/deformable_base_env.py
@@ -1,6 +1,6 @@
 
 from __future__ import annotations
-
+import carb
 import builtins
 import torch
 from ..scene.deformable_interactive_scene import DeformableInteractiveScene
@@ -40,10 +40,21 @@ class DeformableBaseEnv(ManagerBasedEnv):
         print("[INFO]: Base environment:")
         print(f"\tEnvironment device    : {self.device}")
         print(f"\tPhysics step-size     : {self.physics_dt}")
-        print(f"\tRendering step-size   : {self.physics_dt * self.cfg.sim.substeps}")
+        print(f"\tRendering step-size   : {self.physics_dt * self.cfg.sim.render_interval}")
         print(f"\tEnvironment step-size : {self.step_dt}")
         print(f"\tPhysics GPU pipeline  : {self.cfg.sim.use_gpu_pipeline}")
         print(f"\tPhysics GPU simulation: {self.cfg.sim.physx.use_gpu}")
+
+        if self.cfg.sim.render_interval < self.cfg.decimation:
+            msg = (
+                f"The render interval ({self.cfg.sim.render_interval}) is smaller than the decimation "
+                f"({self.cfg.decimation}). Multiple multiple render calls will happen for each environment step. "
+                "If this is not intended, set the render interval to be equal to the decimation."
+            )
+            carb.log_warn(msg)
+
+        # counter for simulation steps
+        self._sim_step_counter = 0
 
         # generate scene
         with Timer("[INFO]: Time taken for scene creation"):