+++ octilab/envs/mdp/actions/__init__.py
@@ -9,4 +9,3 @@ from .actions_cfg import *
 from .binary_joint_actions import *
 from .joint_actions import *
 from .non_holonomic_actions import *
-from .tycho_controller_task_space_actions import *